Allow decompressing test outputs.
This seems to be needed to test fastq outputs from tools. Gzipped files seem to depend on the file name before compression - so diff compares (even binary ones) do not work. This allows comparing compressed outputs (and it will magically work if the expected file in test-data is compressed or not).

Builds on the awesome work from @mvdbeek in #3512.