Remove broken tour step.
This causes the Selenium tests to fail under various conditions - but legimately so. The referenced target link disappears pretty quickly after the step in the previous click. When I run them in the web browser and this step is "disconnected" and in the center of the screen and as a result it isn't clear what links are being referred to.