Add more flexibility to mulled operations.
Clean up verbose handling, make images configurable, and such. The major fix is the addition of allowing a specified version of Conda to be installed before building images - this is required to work around certain packages such ncurses not working under osxfs and conda=4.2. This is fixed with Conda 4.3.

Likely this bug is related to https://github.com/conda/conda/issues/3494.