New Authentication section of the docs.
I pulled in this doc https://galaxyproject.org/develop/authentication/ which was very dated and somewhat incorrect and I filled it out with more details, updated configuration information, and a whole new section on the authentication framework. I also pulled in this doc https://galaxyproject.org/admin/user-accounts/ which was pretty good and I left it largely unchanged.

Rebased to fix index inclusion and with edits suggested by :eagle::eyes:.