Abstraction for re-use between collection creators.
The new mixin reduces the duplication between the list and list of paired collection creators.

I'd like to add functionality related to hiding the original datasets (especially useful for direct upload of collections but previously requested also) but I'd like to reduce the duplication related to doing that first. This is a start in that direction.