Fix multiple collection reductions on multiple data parameters.
Was introduced in #805 but there was a small bug (maybe a last minute change before opening the PR) in the client code so this never was exposed to the users the way it was intended.

This should largely negate the need for hacks such as in cuffmerge that allow multiple list collections in addition to main multiple data selector. xref galaxyproject/tools-iuc#1225.