Add API route for downloading entire collections
Entire collections can now be downloaded as tarballs. I have used the
StreamBall class from galaxy.utils, so archives start downloading
immediately.

The collection structure is mapped onto a directory struture, where each
level of a collection is a directory in the archive.  Collections of BAM
files are included alongside their .bai indices, and composite datatype
are supported as well.

API tests for downloading list, paired and list:paired collections are
included.

Currently collection elements that are not in the OK status will be
skipped, so we should probably hide the download button if not all
collection elements are in a final state.