Add the ability to control job conf plugin loading from environment variables.
All job configuration parameters can be set using environment variables but currently the actual loading of the plugin cannot. This is problematic for setting up single job confs for multiple environemnts (like done with docker-galaxy-stable) when using plugins that may hang waiting for resources (such as Kubernetes for the K8 runner or a MQ for Pulsar for instance) or may require library dependencies not be setup in every environment the job configuration is deployed in.

xref https://github.com/bgruening/docker-galaxy-stable/pull/328

Rebased with advice from @bgruening.