Implement mulled containers v2.
For a given set of packages and versions, the v1 containers had names and versions like:

  mulled-v1-<hash_of_packages_and_versions>:<build_number>

Here the quay.io repository has a unique name per combination of packages and versions - and the repository tag (the part after the :) is solely used to specify a build number (a mechanism allowing us to increment and rebuild containers with same recipes if they are broken in some way - mirroring closely the conda build number).

The version of the hash looks like:

  mulled-v2-<hash_of_packages>:<hash_of_versions>-<build_number>

Here a repository is only created for each combination of packages - different combinations of versions are represented as different tags within that repository.

This commit breaks backward compatibility (of a feature configured with a beta label) - but since we are aware of no biocontainers published in the wild it would probably be a okay idea to eliminate the backward compatability.