Async update 20170615
The asynchronous procedure expects that only one output is defined in
the tool XML schema. This improvement permits to define also
collections. This feature is useful in conjunction with tools like
[galaxy-json-collect-data-source](https://github.com/fabio-cumbo/galaxy-
json-collect-data-source) (please see the README of the tool for more
information).
It does not affect the behaviour of the asynchronous data sources
previously developed.