Add `uninstall_repository` to ToolShedRepositoriesController
And connect routes. It is now possible to POST to
api/tool_shed_repositories to trigger a Repository installation and to
DELETE to api/tool_shed_repositories to uninstall a repository. In
addition one can also DELETE to
api/tool_shed_repositories/<repository_id>, which will uninstall the
repository with repository_id.