Subtract 10 seconds from last update time
This is an ugly hack that seems to be necessary for refreshing the
dataset status in a history to running (=yellow) with some job runners.
The basic local runner does not show this problem. I haven't completely
understood where this issue is coming from. It shouldn't really be a
wrong clock, since in my testing server and client are the same machine.