Uniform indentation of multiline if conditionals
Use 8-spaces indentation for the conditional continuation lines of an
`if`-statement where the conditional part spans multiple lines.

According to PEP-8, using 4-spaces "can produce a visual conflict with
the indented suite of code nested inside the `if`-statement", but the PEP
"takes no explicit position".

There were only 2 places in the whole codebase where 8 spaces were not
used, this will make it uniform.