Add requirement for bcftools to set_metadata tool
The `bcftools` binary is used by the `set_meta()` method of the `Bcf`
datatype class. Having a `<requirement>` for it in the set_metadata tool
makes it possible to use conda to resolve the dependency instead of
forcing the admin to install `bcftools` on the PATH of the Galaxy user.

This will also fix errors like
https://travis-ci.org/galaxyproject/tools-iuc/jobs/266923024#L1994
when running
```
planemo test --conda_dependency_resolution --conda_auto_install
```