Allow updating workflows from partial information
The workflow overview page gets workflow information from the
/api/workflows, which returns a reduced amount of information, most
notably the steps are not being returned. This commit skips updating the
actual workflow steps if the steps are not included in the workflow
payload. We can now update name, tags and whether a workflow should be
shown in the tool panel. We also return owner and number of steps in the
response to importing workflows.