Fix a segfault in uWSGI mules caused by trying to access invalid memory post-fork. There may be more cases like this that we'll have to handle in the mule fixup function.
Context
=======

uWSGI mules don't fork()/exec(), they fork() and then (as documented)
run supplied code[1]. In order to read the URL parameter for IGV builds
in the default tool_data_table_conf.xml, Galaxy imports urlopen() from
six.moves.urllib.request. In Python 2.7, this imports the same function
from urllib2. urllib2 uses (and reuses) a module global "opener"
object[2]. This is initialized during the prefork application startup
when used for the IGV data table, and then the same object is used
postfork for the mule to load the same data table.

It was discovered when attempting to run the mule job handler test under
macOS, it does not occur in testing on Linux, nor does it occur on macOS
with the default tool_conf.xml.sample and datatypes_conf.xml.sample.
Whether or not the segfault was encountered seemed to depend on which
tools or datatypes were loaded, but not in any clear pattern. The cause
is presumably due to differences in memory allocation schemes between
macOS and Linux, and perhaps garbage collection differences depending on
different config file sizes? Regardless, it seems to be a classical case
of holding on to references to memory that has been deallocated
elsewhere.

The segfault is specifically on the socket_getaddrinfo function in the
_socket C module that provides functionality for the socket module in
the standard Python lib. socket.getaddrinfo() is called by urlopen() by
way of httplib.

uWSGI mule debugging
--------------------

I ultimately figured out the cause of the segfault with lldb. To make
that experience better, I used a debug build of both Python and uWSGI
(uWSGI was probably less important to do but that process is explained
here anyway). Python with debugging is not ABI compatible with non-debug
Python, so building all of Galaxy's C Extension wheels was necessary. In
many cases, pip will do this for you (using the standard pip invocation
in scripts/common_startup.sh), but not all will be successful. To build
debug Python and debug uWSGI and run Galaxy under them, I used the
following steps:

1. You will need a newer version OpenSSL than the one that ships with
   Sierra, I used the one from Brew:

   CFLAGS=-I/usr/local/opt/openssl/include \
   LDFLAGS=-L/usr/local/opt/openssl/lib \
   MACOSX_DEPLOYMENT_TARGET=10.12 \
   ./configure --prefix=$HOME/pydebug --with-pydebug

   Be sure to set $MACOSX_DEPLOYMENT_TARGET. If you do not, the
   `getaddrinfo` function will utilize locking around `getaddrinfo` that
   should not be used[3].

2. After intsalling Python, `virtualenv -p ~/pydebug .venv` in the
   Galaxy root.

3. In the uWSGI source, copy buildconf/default.ini to
   buildconf/debug.ini, set `debug = true` in debug.ini and then (with
   .venv activated) run:

   UWSGI_PROFILE=buildconf/debug.ini python setup.py install

4. Manually install a few problematic dependencies:

   - Use the Starforge recipe for building pysam[4], which mainly means
     grabbing a tarball of @dannon's fork, fixing the version number,
     and running `python setup.py install`. The ssdist tarball for this
     verison on wheels.galaxyproject.org is broken (files are missing
     from its manifest) and I need to fix that. Dannon is also currently
     working on upstreaming his changes and updating Galaxy for the
     latest version of pysam so this shouldn't be a problem in the
     future.

   - The mercurial sdist tarball on wheels.galaxyproject.org is also
     broken, download the correct version tarball directly from the
     mercurial release distribution site and `python setup.py install`
     it. I need to fix this as well.

   - Install numpy before bx-python (if it's not installed before bx is,
     bx will be missing needed features):

     pip install -i https://wheels.galaxyproject.org --no-binary :all: \
        $(grep -i numpy= requirements.txt)

5. Install remaining requirements, iirc the rest should be ok:

   pip install -i https://wheels.galaxyproject.org --no-binary :all: \
      -r requirements.txt

   --no-binary :all: is necessary because there are numerous macOS
   wheels in PyPI with a `none` SOABI tag, meaning these wheels will
   install but will not work for the debug Python ABI.

When you run Galaxy under the debugger, you won't be able to directly
control the fork()ed mules. Instead, you'll need to fire up a second
debugger to catch the mule[5]:

1. Start up Galaxy under lldb, set a breakpoint on fork Be sure that
   your virtualenv is *NOT* activated before starting lldb, or it will
   fail miserably at loading its own python module.

   % lldb .venv/bin/uwsgi
   (lldb) b fork

2. Run Galaxy, e.g.:

   (lldb) r --virtualenv /Users/nate/galaxy/.venv --module galaxy.webapps.galaxy.buildapp:uwsgi_app_factory() --yaml config/galaxy.yml --processes 1 --threads 4 --http localhost:8080 --pythonpath lib --master --static-map /static/style=/Users/nate/galaxy/static/style/blue --static-map /static=/Users/nate/galaxy/static --die-on-term --enable-threads --py-call-osafterfork --mule=lib/galaxy/main.py --farm=job-handlers:1

3. The first fork() is for the workers, so you'll want to skip this:

   (lldb) n

4. The second fork is for the mule. At this point, start a second lldb
   in another terminal, set a breakpoint on something interesting,  and
   attach it to the process that is about to start:

   % lldb
   (lldb) b socket_getaddrinfo
   (lldb) pro at -n uwsgi -w

5. Continue the fork in the original lldb:

   (lldb) n

6. Continue debugging in the second lldb, which is now attached to the
   mule.

[1]: https://github.com/unbit/uwsgi/blob/5039a4b679735348422638baa7cff7a8116762e3/core/mule.c#L95
[2]: https://github.com/python/cpython/blob/20958e6d91d11a80d6c664ce6346791259b1d193/Lib/urllib2.py#L130
[3]: https://emptysqua.re/blog/getaddrinfo-deadlock/
[4]: https://github.com/galaxyproject/starforge/blob/f1013ddf6119381ffba6df06e8d996cee5678280/wheels/build/wheels.yml#L49
[5]: https://stackoverflow.com/questions/14746315/how-to-attach-to-child-process-in-lldb