Make submitting logins for Selenium tests a bit more robust.
There are a few different transiently failing tests on Jenkins that seem to fail because the test thinks Galaxy is logged in but it is not. So now after we click the login button in Selenium - we will wait for to see the "Register or Login" button turn into the "User" button.

This is a similar problem and fix to what was done in #4562 - which seemed to help.