Selenium - catch popups fading out as transitions to retry various actions on.
Previously we only detected transitions for retrying actions based on stale element exceptions, this expands that to include popups that may be fading out for instance.

I think this is what is happening with the transiently failing test here https://jenkins.galaxyproject.org/job/selenium/528/artifact/528-test-errors/test_save_as2017092203291506065342/stacktrace.txt. The exception indicates that a click was not clickable because a modal element that was fading out - though it had been previously clickable. I think this should fix that.