Skip recording JobToInputDatasetCollection relations
if the item to record is not a HistoryDatasetCollectionAssociation. This fixes

```
Error executing tool: Attempting to flush an item of type <class
\'galaxy.model.DatasetCollectionElement\'> as a member of collection
"JobToInputDatasetCollectionAssociation.dataset_collection". Expected an object
of type <class \'galaxy.model.HistoryDatasetCollectionAssociation\'> or a
polymorphic subclass of this type.' ```

It isn't ideal, but before 19818cb we didn't record Collection-like elements at all.