Tighten job search result
We now expand incoming tool parameters, so that we have a dictionary of all
tool parameters (i.e specified/posted parameters and defaults). We start the
search by quering the database for jobs using the same tool id for the same
owner with the same input datasets, taking into account copied
dataset(-collection)s.  We then replace all dataset ids (HDA/LDA or HDCA) in
the parameter dict with the input data used for the jobs queried in the first
step. This allows us to check every parameter for a match with the job.