Avoid duplicate dependency installation
The previous commit changed the resolver logic to try all resolvers, so that
when building an environment fails, the next resolver can be tried. Since
dependency resolution and installation are coupled it would be possible to
install dependencies multiple times for different resolvers (e.g for Conda and
Conda versionless).
To avoid this we stop installation once a set of dependencies has been
resolved.