Symlink tbi index to work around pysam limitation
Before https://github.com/pysam-developers/pysam/pull/586 is merged and a new release is out
we create a symlink to the tbi file, which is required for creating TabixFile
instances. Since we want to cleanup the symlinks I turned `get_data_file` into a contextmanager.
Along the way I also changed many open()/close() calls to `with` statements.