Fix job resume functionality for non-prefixed input data
This should fix https://github.com/galaxyproject/galaxy/issues/5222.

The problem was that HDA ids in nested parameters were not always being updated
properly. In the case of bowtie2 the input dataset is provided via a
conditional, but the conditional prefix is not being stored in the
JobToInputDatasetAssociation, and so the `update_param` function was not able
to update the dataset id in these instances.  The approach now is to simply
replace all occurences of the old dataset id with the new dataset id.

TODO: tests, refactor the replacement functionality into a separate function
and make this work for JobToInputDatasetCollectionAssociation.