Wait for dataset to be present when guessing size
I noticed that for very short but numerous grep jobs or very small alignments
I would get datasets that had `empty` written where I would expect the dataset peak,
and alignments would be missing the index file.

I traced this down to the file size being 0, which leads to external metadata
not being read in. The time during which we wait for the dataset is controlled
by the `retry_job_output_collection` setting.