Demonstrate local components broken without template compiler.
The second example of components from the Vue docs (https://vuejs.org/v2/guide/components.html#Local-Registration) seems to not work without the template compiler. I'm okay with us declaring we are not going to use the template compiler at runtime but I want to be clear about what we are loosing - I think it is being able to do this right - defining components in the same file?